Use double for Point components
Last selected hint point had a 67.5 y, which when zoom x8 had a box too short (by 4 pixels).

+++ src/main/java/mousemaster/Point.java
@@ -1,4 +1,4 @@
 package mousemaster;
 
-public record Point(int x, int y) {
+public record Point(double x, double y) {
 }
